api: fan leaderboard, host leaderboard -> green

+++ src/main/java/com/tanvir/features/leaderboard/domain/UserBeanSummary.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UserBeanSummary {
     private String id;
-    private int totalBeans;
+    private double totalBeans;
 }
